requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
numpy>=1.26.0
scikit-learn>=1.4.0
pypdf>=4.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
typer>=0.12.0
rich>=13.7.0
